Refactored to return correct loadingState on complete
